Add new branch for meta-angstrom but "older commit"
This solved two problems one that the ath9k wifi driver crashed, reason was that firmware api (request_firmware call) was patched with somthing that brock the call to get the ath9k firmware. And changing to sumo branch made it possible to use backport 4.18 for ath9k driver again there is a problem to go to "import OE-core 239" but here OE-core 237 seems to work.

Change-Id: I25eef8e5be1c15be5408d73799dc671f59c5cd05